Restructure
- Consolidated houses and gumps in a new Houses folder, inside of Multis.
